fix(win7/virtio-snd): cast NTSTATUS in debug print for /WX
DbgPrintEx is printf-annotated; cast NTSTATUS to UINT when using %08X to avoid format warnings treated as errors.

+++ drivers/windows7/virtio-snd/src/adapter.c
@@ -121,7 +121,7 @@ static NTSTATUS VirtIoSndStartDevice(PDEVICE_OBJECT DeviceObject, PIRP Irp, PRES
 
     status = VirtIoSndHwStart(dx, Irp);
     if (!NT_SUCCESS(status)) {
-        VIRTIOSND_TRACE_ERROR("VirtIoSndHwStart failed: 0x%08X\n", status);
+        VIRTIOSND_TRACE_ERROR("VirtIoSndHwStart failed: 0x%08X\n", (UINT)status);
         VirtIoSndHwStop(dx); // best-effort cleanup of partial allocations
         goto Exit;
     }
